# Improve WebUtility.HtmlEncode / UrlEncode performance